Move casadi_to_lp file
